requirements.txt
requests
schedule
retry
python-dotenv